requirements.txt
fastapi>=0.109
uvicorn>=0.27
pydantic>=2.5,<3
pydantic-settings>=2.1,<3
requests>=2.31
aiohttp>=3.9
redis>=5.0
motor>=3.3
pymongo>=4.6
cachetools>=5.3
python-dotenv>=1.0
pyswisseph>=2.10
apscheduler>=3.10
